getallcountries api added

+++ app/src/main/java/com/example/firozhasan/retrofitkotlinexample/model/modelClass/Country.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Country(
 
 	@field:SerializedName("area")
-	val area: Int? = null,
+	val area: Double? = null,
 
 	@field:SerializedName("nativeName")
 	val nativeName: String? = null,
@@ -72,7 +72,7 @@ data class Country(
 	val region: String? = null,
 
 	@field:SerializedName("latlng")
-	val latlng: List<Int?>? = null,
+	val latlng: List<Double?>? = null,
 
 	@field:SerializedName("currencies")
 	val currencies: List<CurrenciesItem?>? = null
